requirements.txt
absl-py==1.0.0
altgraph==0.17.2
astunparse==1.6.3
cachetools==4.2.4
certifi==2020.6.20
charset-normalizer==2.0.9
colorama==0.4.4
cycler==0.11.0
dataclasses==0.8
Django==2.0
future==0.18.2
gast==0.3.3
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
h5py==2.10.0
idna==3.3
image==1.5.33
importlib-metadata==4.8.2
jupyter-contrib-core==0.3.3
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lmdb==1.2.1
lxml==4.6.3
Markdown==3.3.6
matplotlib==3.3.4
numpy==1.18.5
oauthlib==3.1.1
opencv-python==4.5.4.60
opt-einsum==3.3.0
pandas==1.1.5
pefile==2021.9.3
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.4
PyMuPDF==1.19.3
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
pywin32-ctypes==0.2.0
PyYAML==5.4.1
pyzbar==0.1.8
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scipy==1.4.1
seaborn==0.11.2
six==1.16.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
wincertstore==0.2
wrapt==1.13.3
zipp==3.6.0
zxing==0.14
